detach request logic from operation
